Refactor Order and OrderBook classes to use separate enums
