+++ src/test/java/com/coralblocks/coralme/OrderBookNoSelfTradeTest.java
@@ -1,4 +1,4 @@
-/* 
+/*
  * Copyright 2023 (c) CoralBlocks - http://www.coralblocks.com
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -21,29 +21,29 @@ import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
 
-import com.coralblocks.coralme.Order.CancelReason;
-import com.coralblocks.coralme.Order.ExecuteSide;
-import com.coralblocks.coralme.Order.Side;
-import com.coralblocks.coralme.Order.TimeInForce;
+import com.coralblocks.coralme.CancelReason;
+import com.coralblocks.coralme.ExecuteSide;
+import com.coralblocks.coralme.Side;
+import com.coralblocks.coralme.TimeInForce;
 import com.coralblocks.coralme.util.DoubleUtils;
 
 
 public class OrderBookNoSelfTradeTest {
-	
+
 	private static final long CID_1 = 1005L;
 	private static final long CID_2 = 1006L;
-	
+
 	private OrderBookListener called(OrderBookListener listener, int times) {
 		return Mockito.verify(listener, Mockito.times(times));
 	}
-	
+
 	private void done(OrderBookListener listener) {
 		Mockito.verifyNoMoreInteractions(listener);
 		Mockito.clearInvocations(listener);
 	}
-	
+
 	private static class OrderExecutedCaptor {
-		
+
 		ArgumentCaptor<OrderBook> 		book = ArgumentCaptor.forClass(OrderBook.class);
 		ArgumentCaptor<Long> 			time = ArgumentCaptor.forClass(Long.class);
 		ArgumentCaptor<Order> 			order = ArgumentCaptor.forClass(Order.class);
